pharm_standard_back/project/requirements/common.txt
celery[redis]==4.4.0
kombu==4.6.7
amqp==2.5.2
billiard==3.6.1.0
Django==3.0.2
django-constance[database]==2.6.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
django-polymorphic==2.1.2
django-phonenumber-field==4.0.0
django-redis==4.11.0
django-rest-polymorphic==0.1.8
drf-yasg==1.17.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-extra-fields==2.0.3
cryptography==3.3.1
gunicorn==20.0.4
cryptography==3.3.1
phonenumbers==8.11.2
psycopg2-binary==2.8.4
pytz==2019.3
redis==3.3.11
sentry-sdk==0.14.1
django-lifecycle==0.7.1
openpyxl==3.0.4
lxml==4.5.1
python-dateutil==2.8.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Pillow==8.1.0
requests-ntlm==1.1.0
beautifulsoup4==4.9.3
ipython==7.22.0
saml==0.9.0
signxml==2.8.2
python3-saml==1.10.1
pytest==6.0.1
pytest-django==3.9.0
matplotlib==3.5.3
importlib-metadata==4.13.0
transliterate==1.10.2
